requirements.txt
os
pandas
numpy 
seaborn
google-cloud
-e .